+++ src/org/xjge/ui/Widget2.java
@@ -1,5 +1,10 @@
 package org.xjge.ui;
 
+import java.util.Map;
+import org.xjge.core.Mouse;
+import org.xjge.core.ScreenSplitType;
+import org.xjge.graphics.GLProgram;
+
 /**
  * Created: May 12, 2021
  * 
@@ -8,14 +13,14 @@ package org.xjge.ui;
  */
 public abstract class Widget2 {
     
-    public abstract void update();
+    public abstract void update(double targetDelta, double trueDelta);
     
-    public abstract void render();
+    public abstract void render(Map<String, GLProgram> glPrograms);
     
-    public abstract void relocate();
+    public abstract void relocate(ScreenSplitType splitType, int viewportWidth, int viewportHeight);
     
-    public abstract void processKeyboardInput();
+    public abstract void processKeyboardInput(int key, int action, int mods);
     
-    public abstract void processMouseInput();
+    public abstract void processMouseInput(Mouse mouse);
     
 }